chore: Upgrade demo with import for native Fresco typings chore: Add npmignore fix: Add warrning message if imageUri is set to incorrect value chore: Update tsconfig file according to plugin seed
