+++ demo/references.d.ts
@@ -1,1 +1,2 @@
+/// <reference path="./node_modules/tns-platform-declarations/android.d.ts" />
 /// <reference path="./node_modules/nativescript-fresco/android.d.ts" />
