+++ demo/tsconfig.json
@@ -23,6 +23,7 @@
     },
     "exclude": [
         "node_modules",
+        "../src/node_modules",
         "platforms"
     ]
 }
