+++ src/.npmignore
@@ -1,0 +1,5 @@
+*.map
+*.ts
+!*.d.ts
+tsconfig.json
+sync
