+++ src/nativescript-fresco.android.ts
@@ -305,9 +305,13 @@ export class FrescoDrawee extends commonModule.FrescoDrawee {
                     } else if (this.imageUri.indexOf("/") === 0) {
                         uri = android.net.Uri.parse(`file:${this.imageUri}`);
                     }
-                }
-                if (uri === undefined) {
+                } else {
                     uri = android.net.Uri.parse(this.imageUri);
+                }
+
+                if (!uri) {
+                    console.log(`Error: 'imageUri' not valid: ${this.imageUri}`);
+                    return;
                 }
 
                 let progressiveRenderingEnabledValue = this.progressiveRenderingEnabled !== undefined ? this.progressiveRenderingEnabled : false;
