+++ src/references.d.ts
@@ -1,3 +1,2 @@
-/// <reference path="./android.d.ts" />
-/// <reference path="./node_modules/tns-core-modules/tns-core-modules.d.ts" />
 /// <reference path="./node_modules/tns-platform-declarations/android.d.ts" />
+/// <reference path="./android.d.ts" />
